Misc additions to pattern refactor
